pyproject.toml

[tool.poetry]
name = "anonymous"
version = "0.1.0"
description = ""
authors = ["Martin Højland <molheh@yousee.dk>"]

[tool.poetry.dependencies]
python = "^3.8"
transformers = "^4.5.1"
uvicorn = "^0.13.4"
spacy = "^3.0.5"
pandas = "^1.1.4"
numpy = "^1.19.4"
pydantic = "^1.7.2"
python-dotenv = "^0.15.0"
cachetools = "^4.2.1"
fastapi = "^0.63.0"
torch = "^1.8.1"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}
ipykernel = "^5.3.4"
ipywidgets = "^7.6.3"
jupyterlab = "^2.2.9"
mypy = "^0.770"
isort = "^4.3.21"
autoflake = "^1.3.1"
flake8 = "^3.7.9"
pytest-black = "^0.3.10"
pytest-clarity = "^0.3.0-alpha.0"
pytest-mock = "^3.3.1"
pytest-asyncio = "^0.14.0"
pytest-cov = "^2.11.1"
Faker = "^4.1.1"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--cov --cov-report=xml:cov.xml --black"
testpaths = "tests src"

[tool.coverage.paths]
source = ["src/", "*/site-packages"]

[tool.coverage.report]
show_missing = true
fail_under = 40
exclude_lines = [
  "pragma: no cover",
  "raise NotImplementedError"
]

[tool.black]
line-length = 135
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs         # exclude a few common directories in the
  | \.git          # root of the project
  | \.mypy_cache
  | \.venv
  | \.vscode
  | _build
  | buck-out
  | build
  | dist
  | tmp
)/
'''

[tool.isort]
multi_line_output = 3
line_length = 88
include_trailing_comma = true
